Updated dependency-check plugin
It looks like the old 1.0 feeds are no longer published at this URL?

Either way, we need to update the plugin.  Explicitly setting these URLs was needed in one of the older versions of this plugin after NIST changed something with the feed.  We should be able to use the default now.